fix: IExt 클래스를 __init__.py로 이동 + autoload 익스텐션 제거 방식 변경
