+++ k8s/base/configmaps.yaml
@@ -184,6 +184,7 @@ data:
     /isaac-sim/license.sh && /isaac-sim/privacy.sh && /isaac-sim/isaac-sim.streaming.sh \
         --merge-config="/isaac-sim/config/open_endpoint.toml" \
         --enable omni.clipboard.service \
-        --disable isaacsim.robot_motion.motion_generation.tutorials \
+        --/app/exts/excluded/0=isaacsim.robot_motion.motion_generation.tutorials \
+        --/exts/isaacsim.robot_motion.motion_generation.tutorials/autoload=false \
         $EXTRA_FLAGS \
         "$@"
